Refactor EntityActionsMenu to use the new app service
