+++ assets/src/application/ui/layout/entityActionsMenu/constants.ts
@@ -1,0 +1,1 @@
+export const serviceName = 'entityActions';
