+++ assets/src/application/ui/layout/entityActionsMenu/index.ts
@@ -1,7 +1,9 @@
 export { default as EntityActionsMenu } from './EntityActionsMenu';
 
-export { default as useEntityActions } from './useEntityActions';
+export { default as useEntityActionsService } from './useEntityActionsService';
 
-export { default as useEntityActionsManager } from './useEntityActionsManager';
+export { default as useEntityActionsMenu } from './useEntityActionsMenu';
 
 export * from './types';
+
+export * from './constants';
