+++ assets/src/application/ui/layout/entityActionsMenu/types.ts
@@ -1,5 +1,48 @@
 import React from 'react';
 import { Entity } from '@appServices/apollo/types';
+import { Subscriptions } from '@appServices/subscription';
+
+export interface SubscriptionsOptions<T extends string> {
+	entityType?: T; // to limit the subscription only to specific entityType
+}
+
+export interface SubscriptionData<E extends Entity, T extends string> extends SubscriptionsOptions<T> {
+	entity: E;
+}
+
+export interface EntityActionsService {
+	subscribe: SubscribeFn;
+	getSubscriptions: <E extends Entity, T extends string>(
+		options?: SubscriptionsOptions<T>
+	) => Subscriptions<SubscriptionData<E, T>, SubscriptionsOptions<T>>;
+}
+
+export type EntityActionsServiceHook = <Domain extends string>(domain: Domain) => EntityActionsService;
+
+export type SubscribeFn = <E extends Entity, T extends string>(
+	cb: EntityActionsSubscriptionCb<E, T>,
+	options?: SubscriptionsOptions<T>
+) => VoidFunction;
+
+export type EntityActionsSubscriptionCb<E extends Entity, T extends string> = (
+	data: SubscriptionData<E, T>,
+	entityActionsMenu: EntityActionsMenu
+) => void;
+
+export interface EntityActionsMenu {
+	registerMenuItem: (key: string, component: React.ReactType) => void;
+	unRegisterMenuItem: (key: string) => void;
+	getMenuItems: () => EntityMenuItems;
+}
+
+export interface ActionsMenuComponentProps<E extends Entity> {
+	entity: E;
+	[key: string]: any;
+}
+
+export type EntityMenuItems = {
+	[key: string]: React.ReactType;
+};
 
 /**
  * e.g.
@@ -25,63 +68,3 @@ export type MenuRegistry = {
 		[key: string]: EntityMenuItems;
 	};
 };
-
-export type SubscribeFn = <E extends Entity, T extends string>(
-	cb: SubscriptionCallback<E, T>,
-	options?: SubscriptionsOptions<T>
-) => VoidFunction;
-
-export type SubscriptionCallback<E extends Entity, T extends string> = (
-	data: SubscriptionData<E, T>,
-	entityActionsManager: EntityActionsManager
-) => void;
-
-export interface SubscriptionData<E, T> {
-	entityType: T;
-	entity: E;
-}
-
-export interface EntityActionsManager {
-	registerMenuItem: (key: string, component: React.ReactType) => void;
-	unRegisterMenuItem: (key: string) => void;
-	getMenuItems: () => EntityMenuItems;
-}
-
-export type EntityMenuItems = {
-	[key: string]: React.ReactType;
-};
-
-export interface SubscriptionsOptions<T extends string> {
-	entityType?: T; // to limit the subscription only to specific entityType
-}
-
-export interface Subscription<E extends Entity, T extends string> {
-	callback: SubscriptionCallback<E, T>;
-	options: SubscriptionsOptions<T>;
-}
-
-export type Subscriptions<E extends Entity, T extends string> = {
-	[key: string]: Subscription<E, T>;
-};
-
-export interface EntityActionsData {
-	subscribe: SubscribeFn;
-	subscriptions: Subscriptions<any, string>;
-}
-
-export interface EntityActions {
-	subscribe: SubscribeFn;
-	getSubscriptions: <E extends Entity, T extends string>(options?: SubscriptionsOptions<T>) => Subscriptions<E, T>;
-}
-
-export interface UpdateSubscriptionProps {
-	id: string;
-	callback?: SubscriptionCallback<any, string>;
-	options?: SubscriptionsOptions<string>;
-	action?: 'add' | 'remove';
-}
-
-export interface ActionsMenuComponentProps<E extends Entity> {
-	entity: E;
-	[key: string]: any;
-}
